Added a summarize function + minor tweaks
Note: There is some variation in the performance of glmnet/fista/BPDN for small values of (m,n)... check that b \in Im(A) in the future? Stability issues?

Next: Follow up on this AND attempt to speed up the code with external QR updates... this is really the main issue...